fix CI publish
- change CI publish to trigger on released release

Git tags can only be created in github by creating a release. So instead
of triggering on a new tag push, just trigger on a release being
released. This will end up just adding the assets to the release after
the workflow finishes.

- save name of asset before uploading

Because upload-release-asset does not do glob expansion, we have to
save the full file name in a step before we do the upload.

- add write permission for publishing assets

-----
change CI Build trigger

- Remove triggering on push
- Add triggering on pull request

Push triggers when tags are pushed and this isn't what we want. Instead
we want to trigger on pull request since that is when the repo is
most volitile anyway. We still have the option to trigger the workflow
manually.