use dependency injection to get Repository
For later testing, we'll need to be able to specify the location
of a database other than the default location. To do this, we have
to provide Repository with the location. To do this we have to create
some helper interfaces that will allow us to inject the location
for normal usage and allow tests to provide their own locations.