refactor formatting display of dates
Using a proptery to expose StartDate and RecordedDate as text
was a workaround for not knowing how to get MAUI XAML to format dates
as strings. I figured out how to do that so we don't need those
specific text properties.

Also, those text properties would make converting to EntityFramework
hard.